requirements.txt
selenium
tqdm
colorama
spotipy
python-dotenv
pytest